Refactor logic for updating daily scripture reveal
